+++ packages/proto-signing/src/pubkey.spec.ts
@@ -8,6 +8,9 @@ describe("pubkey", () => {
   const defaultPubkeyBase64 = "AtQaCqFnshaZQp6rIkvAPyzThvCvXSDO+9AzbxVErqJP";
   const defaultPubkeyBytes = fromBase64(defaultPubkeyBase64);
   const defaultPubkeyProtoBytes = Uint8Array.from([0x0a, defaultPubkeyBytes.length, ...defaultPubkeyBytes]);
+  const ed25519PubkeyBase64 = "kEX3edqZB+HdCV92TPS7ePX0DtP62GWIjmrveZ5pnaQ=";
+  const ed25519PubkeyBytes = fromBase64(ed25519PubkeyBase64);
+  const ed25519PubkeyProtoBytes = Uint8Array.from([0x0a, ed25519PubkeyBytes.length, ...ed25519PubkeyBytes]);
 
   describe("encodePubkey", () => {
     it("works for secp256k1", () => {
@@ -38,6 +41,17 @@ describe("pubkey", () => {
       expect(decodePubkey(pubkey)).toEqual({
         type: "tendermint/PubKeySecp256k1",
         value: defaultPubkeyBase64,
+      });
+    });
+
+    it("works for ed25519", () => {
+      const pubkey = {
+        typeUrl: "/cosmos.crypto.ed25519.PubKey",
+        value: ed25519PubkeyProtoBytes,
+      };
+      expect(decodePubkey(pubkey)).toEqual({
+        type: "tendermint/PubKeyEd25519",
+        value: ed25519PubkeyBase64,
       });
     });